First pass at using new A6 external tags interface
Most of the code is to move the check to use the Agent toolkit.

* use pip-tools to track the reqfile
* make the vsphere check use the toolkit
* added vsphere to the checks supporting the toolkit
* fixed call to add_external_tags(), add tests
* call add_external_tags() with the output from get_external_host_tags()
* fixed according to A6 implementation
* bumped version
