+++ vsphere/CHANGELOG.md
@@ -1,37 +1,26 @@
 # CHANGELOG - vsphere
 
-1.0.4 / 2017-10-10
-==================
+## 2.0.0 / Unreleased
 
-### Changes
+ * [FEATURE] Run with Agent versions >= 6.
+
+## 1.0.4 / 2017-10-10
 
 * [BUGFIX] Fix a possible leak of the vSphere password in the collector logs. See [#722][].
 
-1.0.3 / 2017-08-28
-==================
-
-### Changes
+## 1.0.3 / 2017-08-28
 
 * [BUGFIX] Fix case where metrics metadata don't contain what we expect.
 
-1.0.2 / 2017-07-18
-==================
-
-### Changes
+## 1.0.2 / 2017-07-18
 
 * [SANITY] Import `Timer` helper from `utils.timer` instead of deprecated `util`. See [#484][]
 
-1.0.1 / 2017-06-05
-==================
-
-### Changes
+## 1.0.1 / 2017-06-05
 
 * [BUGFIX] Fix case where returned data series are empty. See [#346][]
 
-1.0.0 / 2017-03-22
-==================
-
-### Changes
+## 1.0.0 / 2017-03-22
 
 * [FEATURE] adds vsphere integration.
