+++ vsphere/datadog_checks/vsphere/__about__.py
@@ -1,4 +1,5 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-__path__ = __import__('pkgutil').extend_path(__path__, __name__)
+
+__version__ = "2.0.0"
