+++ vsphere/datadog_checks/vsphere/common.py
@@ -1,0 +1,4 @@
+# (C) Datadog, Inc. 2010-2017
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+SOURCE_TYPE = 'vsphere'
