+++ vsphere/manifest.json
@@ -11,7 +11,6 @@
     "mac_os",
     "windows"
   ],
-  "version": "1.0.4",
   "guid": "930b1839-cc1f-4e7a-b706-0e8cf3218464",
   "public_title": "Datadog-vSphere Integration",
   "categories":["cloud"],
