+++ vsphere/setup.cfg
@@ -1,0 +1,9 @@
+[aliases]
+test = pytest
+
+[tool:pytest]
+testpaths = ./tests
+
+[flake8]
+exclude = .eggs,.tox
+max-line-length = 120
