+++ vsphere/setup.py
@@ -1,74 +1,34 @@
-# Always prefer setuptools over distutils
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
 from setuptools import setup
-# To use a consistent encoding
 from codecs import open
 from os import path
 
-import json
-import re
+HERE = path.abspath(path.dirname(__file__))
 
-here = path.abspath(path.dirname(__file__))
-
-def parse_req_line(line):
-    line = line.strip()
-    if not line or line.startswith('--hash') or line[0] == '#':
-        return None
-    req = line.rpartition('#')
-    if len(req[1]) == 0:
-        line = req[2].strip()
-    else:
-        line = req[1].strip()
-
-    if '--hash=' in line:
-        line = line[:line.find('--hash=')].strip()
-    if ';' in line:
-        line = line[:line.find(';')].strip()
-    if '\\' in line:
-        line = line[:line.find('\\')].strip()
-
-    return line
+# Get version info
+ABOUT = {}
+with open(path.join(HERE, "datadog_checks", "vsphere", "__about__.py")) as f:
+    exec(f.read(), ABOUT)
 
 # Get the long description from the README file
-with open(path.join(here, 'README.md'), encoding='utf-8') as f:
+with open(path.join(HERE, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
+
 # Parse requirements
-runtime_reqs = ['datadog-checks-base']
-with open(path.join(here, 'requirements.txt'), encoding='utf-8') as f:
-    for line in f.readlines():
-        req = parse_req_line(line)
-        if req:
-            runtime_reqs.append(req)
+def get_requirements(fpath):
+    with open(path.join(HERE, fpath), encoding='utf-8') as f:
+        return f.readlines()
 
-def read(*parts):
-    with open(path.join(here, *parts), 'r') as fp:
-        return fp.read()
-
-def find_version(*file_paths):
-    version_file = read(*file_paths)
-    version_match = re.search(r"^__version__ = ['\"]([^'\"]*)['\"]",
-                              version_file, re.M)
-    if version_match:
-        return version_match.group(1)
-    raise RuntimeError("Unable to find version string.")
-
-# https://packaging.python.org/guides/single-sourcing-package-version/
-version = find_version("datadog_checks", "vsphere", "__init__.py")
-
-manifest_version = None
-with open(path.join(here, 'manifest.json'), encoding='utf-8') as f:
-    manifest = json.load(f)
-    manifest_version = manifest.get('version')
-
-if version != manifest_version:
-    raise Exception("Inconsistent versioning in module and manifest - aborting wheel build")
 
 setup(
     name='datadog-vsphere',
-    version=version,
-    description='The Vsphere check',
+    version=ABOUT["__version__"],
+    description='The vSphere check',
     long_description=long_description,
-    keywords='datadog agent vsphere check',
+    keywords='datadog agent vSphere check',
 
     # The project's main homepage.
     url='https://github.com/DataDog/integrations-core',
@@ -86,7 +46,7 @@ setup(
         'Intended Audience :: Developers',
         'Intended Audience :: System Administrators',
         'Topic :: System :: Monitoring',
-        'License :: OSI Approved :: MIT License',
+        'License :: OSI Approved :: BSD License',
         'Programming Language :: Python :: 2',
         'Programming Language :: Python :: 2.7',
     ],
@@ -95,26 +55,23 @@ setup(
     packages=['datadog_checks.vsphere'],
 
     # Run-time dependencies
-    install_requires=list(set(runtime_reqs)),
+    install_requires=get_requirements('requirements.in')+[
+        'datadog-checks-base',
+    ],
 
     # Development dependencies, run with:
     # $ pip install -e .[dev]
     extras_require={
         'dev': [
             'check-manifest',
-            'datadog_agent_tk>=5.15',
         ],
     },
 
     # Testing setup and dependencies
-    tests_require=[
-        'nose',
-        'coverage',
-        'datadog_agent_tk>=5.15',
-    ],
-    test_suite='nose.collector',
+    setup_requires=['pytest-runner', ],
+    tests_require=get_requirements(path.join('tests', 'requirements.txt')),
 
     # Extra files to ship with the wheel package
-    package_data={b'datadog_checks.vsphere': ['conf.yaml.example']},
+    package_data={'datadog_checks.vsphere': ['conf.yaml.default']},
     include_package_data=True,
 )
