+++ vsphere/tests/fixtures/vsphere_topology.json
