+++ vsphere/tests/requirements.txt
@@ -1,0 +1,2 @@
+mock==2.0.0
+pytest
