+++ vsphere/tox.ini
@@ -1,0 +1,22 @@
+[tox]
+minversion = 2.0
+basepython = py27
+envlist =
+    vsphere
+    flake8
+
+[testenv]
+platform = linux2|darwin
+
+[testenv:vsphere]
+deps =
+    ../datadog-checks-base
+    -rtests/requirements.txt
+commands =
+    pip install --require-hashes -r requirements.txt
+    pytest
+
+[testenv:flake8]
+skip_install = true
+deps = flake8
+commands = flake8 .
